backend/pyproject.toml
[project]
name = "partselect-agent"
version = "0.1.0"
description = "PartSelect chat agent backend — FastAPI + Claude tool use + Chroma."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.1",
    "anthropic>=0.40",
    "openai>=1.54",
    "chromadb>=0.5",
    "httpx>=0.27",
    "curl-cffi>=0.7",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"
